Remove useImageUpload hook and add the functions to ImageUploadUtils
